made server-side security updates

+++ app/controllers/api/v1/auditions_controller.rb
@@ -24,31 +24,42 @@ before_action :require_login
 
     audition = Audition.find(params[:id])
     gig_request = GigRequest.find(params[:gig_request_id])
-    video_id = audition.video_id
+    if current_user == gig_request.user
+        video_id = audition.video_id
 
-    url = URI("https://api.addpipe.com/video/#{video_id}")
+        url = URI("https://api.addpipe.com/video/#{video_id}")
 
-    http = Net::HTTP.new(url.host, url.port)
-    http.use_ssl = true
-    request = Net::HTTP::Delete.new(url)
-    request["x-pipe-auth"] = ENV['PIPE_API_KEY']
-    request["content-type"] = 'application/json'
-    request["cache-control"] = 'no-cache'
+        http = Net::HTTP.new(url.host, url.port)
+        http.use_ssl = true
+        request = Net::HTTP::Delete.new(url)
+        request["x-pipe-auth"] = ENV['PIPE_API_KEY']
+        request["content-type"] = 'application/json'
+        request["cache-control"] = 'no-cache'
 
-    response = http.request(request)
-    puts response.read_body
+        response = http.request(request)
+        puts response.read_body
 
-    audition.destroy
+        audition.destroy
 
-    audition_data = gig_request.auditions.pluck(:id, :name)
-    gig_auditions = audition_data.map do |audition|
-      this_audition = Audition.find(audition[0])
-      audition_user = this_audition.user
-      audition << audition_user
+        audition_data = gig_request.auditions.pluck(:id, :name)
+        gig_auditions = audition_data.map do |audition|
+          this_audition = Audition.find(audition[0])
+          audition_user = this_audition.user
+          audition << audition_user
+        end
+        flash[:alert] = "Audition Deleted"
+        render json: gig_auditions
+      end
+    else
+      flash.now[:alert] = 'Unauthorized User. Could not delete record.'
+      audition_data = gig_request.auditions.pluck(:id, :name)
+      gig_auditions = audition_data.map do |audition|
+        this_audition = Audition.find(audition[0])
+        audition_user = this_audition.user
+        audition << audition_user
+      end
+      render json: gig_auditions
     end
-    flash.now[:notice] = "Audition Deleted"
-    render json: gig_auditions
-  end
 
   private
   def gig_req_audition_params
